Cleanup and rename to vidstream
Deleted all the files that are not needed and were just template junk introduced by creating the project through Xcode with the default stuffs.

Renamed everything from vidtest to vidstream, as the naming was only due to the initial creating being a test. It is no longer a test so the name needed fixing for a while now.

Changed the visible name of the app ( as an icon and within the casting selecting ) to be "CF Vidstream" to reflect its purpose of streaming video to ControlFloor.