Add logging port. Improve performance.
Added a nng port to stream some logs out since they weren't workign through the normal Apple logging methods.
The Apple methods seemed to be causing the extension to crash.

Tweaked image comparison for still frame detection to only compare a single plane and to exit early once the
image is clearly different.

Reduced the output vertical height of video to 848 from 1000, and reduced the quality of jpeg to 0.8.