Fix activation via app on some devices
Some devices were not functioning allowing activation through the app due to an Apple bug.
The bug is that UIScene lifecycle is not compatible with the broadcast picker on some devices.
Changed to AppDelegate lifecycle to make it work. This will also allow the app to work on older iOS versions.

Fixed all the various warnings in the ujsonin code. Just got tired of seeing them.